--- a/SigColorFastAviUtl.sln
+++ b/SigColorFastAviUtl.sln
@@ -5,16 +5,30 @@ VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "SigColorFastAviUtl", "SigColorFastAviUtl\SigColorFastAviUtl.vcxproj", "{12C3C998-F502-4B2E-9A79-AF67D63094A5}"
 EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "test", "test\test.vcxproj", "{22BD0CC7-1611-441A-86BB-93F3E541E393}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
 		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{12C3C998-F502-4B2E-9A79-AF67D63094A5}.Debug|x64.ActiveCfg = Debug|Win32
 		{12C3C998-F502-4B2E-9A79-AF67D63094A5}.Debug|x86.ActiveCfg = Debug|Win32
 		{12C3C998-F502-4B2E-9A79-AF67D63094A5}.Debug|x86.Build.0 = Debug|Win32
+		{12C3C998-F502-4B2E-9A79-AF67D63094A5}.Release|x64.ActiveCfg = Release|Win32
 		{12C3C998-F502-4B2E-9A79-AF67D63094A5}.Release|x86.ActiveCfg = Release|Win32
 		{12C3C998-F502-4B2E-9A79-AF67D63094A5}.Release|x86.Build.0 = Release|Win32
+		{22BD0CC7-1611-441A-86BB-93F3E541E393}.Debug|x64.ActiveCfg = Debug|x64
+		{22BD0CC7-1611-441A-86BB-93F3E541E393}.Debug|x64.Build.0 = Debug|x64
+		{22BD0CC7-1611-441A-86BB-93F3E541E393}.Debug|x86.ActiveCfg = Debug|Win32
+		{22BD0CC7-1611-441A-86BB-93F3E541E393}.Debug|x86.Build.0 = Debug|Win32
+		{22BD0CC7-1611-441A-86BB-93F3E541E393}.Release|x64.ActiveCfg = Release|x64
+		{22BD0CC7-1611-441A-86BB-93F3E541E393}.Release|x64.Build.0 = Release|x64
+		{22BD0CC7-1611-441A-86BB-93F3E541E393}.Release|x86.ActiveCfg = Release|Win32
+		{22BD0CC7-1611-441A-86BB-93F3E541E393}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE