--- a/SigColorFastAviUtl/Histogram.cpp
+++ b/SigColorFastAviUtl/Histogram.cpp
@@ -1,4 +1,4 @@
-#include "Histogram.h"
+﻿#include "Histogram.h"
 #include <exception>
 #include <stdexcept>