--- a/SigColorFastAviUtl/LUT.h
+++ b/SigColorFastAviUtl/LUT.h
@@ -1,4 +1,4 @@
-#pragma once
+﻿#pragma once
 #include <map>
 #include <stdexcept>
 #include <exception>