--- a/SigColorFastAviUtl/RSigmoidTable.h
+++ b/SigColorFastAviUtl/RSigmoidTable.h
@@ -1,10 +1,47 @@
 ﻿#pragma once
-#include "LUT.h"
-class RSigmoidTable :
-	public LUT
+#include <cstdint>
+#include <array>
+#include "sigmoid.hpp"
+#include "thread.hpp"
+class RSigmoidTable
 {
 public:
-	RSigmoidTable(float midtone, float strength, int bin, float multiplier);
-	~RSigmoidTable();
+	using value_type = std::uint16_t;
+	static constexpr value_type table_size = 4097;
+	static constexpr value_type bin = table_size - 1;
+	static constexpr float multiplier = bin;
+private:
+	std::array<value_type, table_size> table_;
+	float midtone_;
+	float strength_;
+public:
+	RSigmoidTable() = default;
+	RSigmoidTable(const RSigmoidTable&) = delete;
+	RSigmoidTable(RSigmoidTable&&) = delete;
+	RSigmoidTable& operator=(const RSigmoidTable&) = delete;
+	RSigmoidTable& operator=(RSigmoidTable&&) = delete;
+	void change_param(float midtone, float strength) noexcept;
+
+	template<typename Unsigned, std::enable_if_t<std::is_unsigned<Unsigned>::value && (sizeof(value_type) <= sizeof(Unsigned)), std::nullptr_t> = nullptr>
+	value_type lookup(Unsigned key) const noexcept { return (table_size <= key) ? table_.back() : table_[key]; }
+
+	template<typename T, std::enable_if_t<
+		(std::is_signed<T>::value && (sizeof(value_type) <= sizeof(T))) || std::is_floating_point<T>::value, 
+	std::nullptr_t> = nullptr>
+	value_type lookup(T key) const noexcept { return (key < 0 || static_cast<T>(table_size) <= key) ? table_.back() : table_[static_cast<value_type>(key)]; }
 };
+inline void RSigmoidTable::change_param(float midtone, float strength) noexcept
+{
+	//0.0 <= midtone <= 1.0, 1.0 <= strength <= 30.0
+	if (midtone == this->midtone_ && strength == this->strength_) return;
+	value_type pre = table_.front() = 0;
+	for (value_type y = 1; y < bin; y++)
+	{
+		const auto x = static_cast<value_type>(multiplier * sigmoid(midtone, strength, static_cast<float>(y) / multiplier));
+		for (value_type i = pre + 1; i < x; ++i) table_[i] = y - 1;//fill blanc
+		table_[x] = y;
+		pre = x;
+	}
+	table_.back() = bin;
+}