--- a/SigColorFastAviUtl/RSigmoidTable.h
+++ b/SigColorFastAviUtl/RSigmoidTable.h
@@ -1,6 +1,7 @@
 ﻿#pragma once
 #include <cstdint>
 #include <array>
+#include <algorithm>
 #include "sigmoid.hpp"
 #include "thread.hpp"
 class RSigmoidTable
@@ -14,6 +15,7 @@ private:
 	std::array<value_type, table_size> table_;
 	float midtone_;
 	float strength_;
+	void insert_to_table(std::size_t pos, value_type value) noexcept;
 public:
 	RSigmoidTable() = default;
 	RSigmoidTable(const RSigmoidTable&) = delete;
@@ -30,18 +32,22 @@ public:
 	std::nullptr_t> = nullptr>
 	value_type lookup(T key) const noexcept { return (key < 0 || static_cast<T>(table_size) <= key) ? table_.back() : table_[static_cast<value_type>(key)]; }
 };
+inline void RSigmoidTable::insert_to_table(std::size_t pos, value_type value) noexcept
+{
+	//to avoid multiple insertation, check target element whether this is first insert.
+	if (0 != pos && 0 == table_[pos]) table_[pos] = value;
+}
 inline void RSigmoidTable::change_param(float midtone, float strength) noexcept
 {
 	//0.0 <= midtone <= 1.0, 1.0 <= strength <= 30.0
 	if (midtone == this->midtone_ && strength == this->strength_) return;
-	value_type pre = table_.front() = 0;
-	for (value_type y = 1; y < bin; y++)
+	//clear array for checking multiple insertation at function `RSigmoidTable::insert_to_table`
+	std::fill(table_.begin(), table_.end(), 0);
+	for (value_type pre = 0, y = 1; y <= bin; ++y)
 	{
 		const auto x = static_cast<value_type>(multiplier * sigmoid(midtone, strength, static_cast<float>(y) / multiplier));
-		for (value_type i = pre + 1; i < x; ++i) table_[i] = y - 1;//fill blanc
-		table_[x] = y;
+		for (value_type i = pre + 1; i <= x; ++i) this->insert_to_table(i, y);//fill blanc and insert new value
 		pre = x;
 	}
-	table_.back() = bin;
 }