--- a/SigColorFastAviUtl/SigColorFastAviUtl.vcxproj
+++ b/SigColorFastAviUtl/SigColorFastAviUtl.vcxproj
@@ -82,7 +82,6 @@
       <FavorSizeOrSpeed>Speed</FavorSizeOrSpeed>
       <EnableEnhancedInstructionSet>StreamingSIMDExtensions2</EnableEnhancedInstructionSet>
       <FloatingPointModel>Fast</FloatingPointModel>
-      <EnableParallelCodeGeneration>true</EnableParallelCodeGeneration>
       <AdditionalOptions>/Qpar-report:1 %(AdditionalOptions)</AdditionalOptions>
       <Parallelization>false</Parallelization>
       <GenerateAlternateCodePaths>None</GenerateAlternateCodePaths>