--- a/SigColorFastAviUtl/SigColorFastAviUtl.vcxproj
+++ b/SigColorFastAviUtl/SigColorFastAviUtl.vcxproj
@@ -105,6 +105,7 @@
     <ClInclude Include="LUT.h" />
     <ClInclude Include="RSigmoidTable.h" />
     <ClInclude Include="SigmoidTable.h" />
+    <ClInclude Include="thread.hpp" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Histogram.cpp" />