--- a/SigColorFastAviUtl/SigColorFastAviUtl.vcxproj
+++ b/SigColorFastAviUtl/SigColorFastAviUtl.vcxproj
@@ -104,6 +104,7 @@
     <ClInclude Include="Histogram.h" />
     <ClInclude Include="LUT.h" />
     <ClInclude Include="RSigmoidTable.h" />
+    <ClInclude Include="sigmoid.hpp" />
     <ClInclude Include="SigmoidTable.h" />
     <ClInclude Include="thread.hpp" />
   </ItemGroup>
@@ -112,7 +113,6 @@
     <ClCompile Include="LUT.cpp" />
     <ClCompile Include="RSigmoidTable.cpp" />
     <ClCompile Include="sigcolorfastaviutl.cpp" />
-    <ClCompile Include="SigmoidTable.cpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">