--- a/SigColorFastAviUtl/SigColorFastAviUtl.vcxproj
+++ b/SigColorFastAviUtl/SigColorFastAviUtl.vcxproj
@@ -101,15 +101,12 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClInclude Include="filter.h" />
-    <ClInclude Include="LUT.h" />
     <ClInclude Include="RSigmoidTable.h" />
     <ClInclude Include="sigmoid.hpp" />
     <ClInclude Include="SigmoidTable.h" />
     <ClInclude Include="thread.hpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="LUT.cpp" />
-    <ClCompile Include="RSigmoidTable.cpp" />
     <ClCompile Include="sigcolorfastaviutl.cpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />