--- a/SigColorFastAviUtl/SigColorFastAviUtl.vcxproj.filters
+++ b/SigColorFastAviUtl/SigColorFastAviUtl.vcxproj.filters
@@ -33,6 +33,9 @@
     <ClInclude Include="thread.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="sigmoid.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Histogram.cpp">
@@ -42,9 +45,6 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="sigcolorfastaviutl.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="SigmoidTable.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="RSigmoidTable.cpp">