--- a/SigColorFastAviUtl/SigColorFastAviUtl.vcxproj.filters
+++ b/SigColorFastAviUtl/SigColorFastAviUtl.vcxproj.filters
@@ -18,9 +18,6 @@
     <ClInclude Include="filter.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="LUT.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
     <ClInclude Include="SigmoidTable.h">
       <Filter>Header Files</Filter>
     </ClInclude>
@@ -35,13 +32,7 @@
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="LUT.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
     <ClCompile Include="sigcolorfastaviutl.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="RSigmoidTable.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>