--- a/SigColorFastAviUtl/SigmoidTable.h
+++ b/SigColorFastAviUtl/SigmoidTable.h
@@ -2,6 +2,7 @@
 #include <cstdint>
 #include <array>
 #include "sigmoid.hpp"
+#include "thread.hpp"
 class SigmoidTable
 {
 public:
@@ -34,10 +35,12 @@ inline void SigmoidTable::change_param(float midtone, float strength) noexcept
 	//0.0 <= midtone <= 1.0, 1.0 <= strength <= 30.0
 	if (midtone == this->midtone_ && strength == this->strength_) return;
 	table_.front() = 0;
-	for (value_type x = 1; x < bin; x++)
-	{
-		table_[x] = static_cast<value_type>(multiplier * sigmoid(midtone, strength, static_cast<float>(x) / multiplier));
-	}
+	parallel::par_for<value_type>(1, bin, [this](value_type begin, value_type end, float midtone, float strength) {
+		for (value_type x = begin; x < end; x++)
+		{
+			table_[x] = static_cast<value_type>(multiplier * sigmoid(midtone, strength, static_cast<float>(x) / multiplier));
+		}
+	}, midtone, strength);
 	table_.back() = bin;
 }