--- a/SigColorFastAviUtl/sigcolorfastaviutl.cpp
+++ b/SigColorFastAviUtl/sigcolorfastaviutl.cpp
@@ -3,9 +3,7 @@
 #include "filter.h" //please set this to AviUtl SDK's filter.h
 #include "SigmoidTable.h"
 #include "RSigmoidTable.h"
-#include <thread>
-#include <type_traits>
-#include <vector>
+#include "thread.hpp"
 #define USECLOCK
 #ifdef USECLOCK
 #include <chrono>
@@ -173,38 +171,6 @@ BOOL func_init_con(FILTER *fp)
 BOOL func_init_sd(FILTER *fp)
 {
 	return TRUE;
-}
-
-namespace parallel {
-	template<
-		typename Index, typename Func, 
-		typename RawIndexType = Index, 
-		std::enable_if_t<std::is_unsigned<Index>::value && std::is_integral<RawIndexType>::value, std::nullptr_t> = nullptr
-	>
-	void par_for(Index num, Func&& f) {
-		const unsigned int thread_num = std::thread::hardware_concurrency();
-		if (thread_num < 2) {//thread 非対応
-			f(static_cast<RawIndexType>(0), static_cast<RawIndexType>(num));
-		}
-		else {
-			const auto task_num = num / thread_num;
-			const auto task_rest = num % thread_num;
-			std::vector<std::thread> th;
-			th.reserve(thread_num);
-			for (unsigned int i = 0; i < thread_num; ++i) {
-				th.emplace_back(
-					std::forward<Func>(f),
-					static_cast<RawIndexType>((i) ? i * task_num + task_rest : 0),
-					static_cast<RawIndexType>((i + 1) * task_num + task_rest)
-				);
-			}
-			for (auto&& t : th) t.join();
-		}
-	}
-	template<typename Index, typename Func, std::enable_if_t<std::is_signed<Index>::value, std::nullptr_t> = nullptr>
-	void par_for(Index num, Func&& f) {
-		if (0 < num) par_for<std::uintmax_t, Func, Index, nullptr>(static_cast<std::uintmax_t>(num), std::forward<Func>(f));
-	}
 }
 namespace color_cvt {
 	inline void yc2rgb(float(&buf)[4], const PIXEL_YC* px) noexcept {