… SigColorFastAviUtl/sigmoid.hpp
… SigColorFastAviUtl/thread.hpp
… test/LUT.cpp
… test/LUT.h
… test/RSigmoidTable.cpp
… test/RSigmoidTable.h
… test/SigmoidTable.cpp
… test/SigmoidTable.h
… test/no_min_max.h
… test/test.cpp
… test/test.vcxproj
… test/test.vcxproj.filters