use `std::thread`.
# benchmark

```
960*720
SDeContrast
Midtone : 50
Strength: 6
```

```
max:最大値
min:最小値
sum:合計
avg:平均
stdev:母集団標準偏差
se:標準誤差
95%CL:95%信頼区間
C.L.max:95%信頼区間を平均に加えたもの
C.L.min:95%信頼区間から平均を引いたもの
```

|  | before[ns] | after[ns] |
| --- | --- | --- |
| max | 70144353 | 90190587 |
| min | 25787378 | 22779376 |
| avg. | 38799114.8973277 | 36900188.2022989 |
| count | 2133 | 2175 |
| stdev | 4120405.38657474 | 8409477.70072626 |
| se | 89216.3637673064 | 180318.129087751 |
| 95%CL | 174860.859815545 | 353417.038771636 |
| C.L.max | 38973975.7571433 | 37253605.2410705 |
| C.L.min | 38624254.0375122 | 36546771.1635272 |