@MaverickTse 2016-10-30
Sorry, nope. The change is too big for such marginal improvement. If there are 10ms difference on 720P, then I'll consider.