@MaverickTse 2016-10-30
Also you need to turn off the /Qpar option in the project file, otherwise the results cannot be compared.