@MaverickTse 2016-10-30
I'd suggest against working on this further more. I have checked on my computer that /Qpar works correctly(using 12 out of my 13 cpu cores) and now I use a simple float[4] array on stack which posed no problem on row-wise parallelization.
The use of std::thread really makes the code a lot more complex and I have no confident I can maintain that.