@yumetodo 2016-10-31
|          | before[ns]       | after[ns]        |
|----------|------------------|------------------|
| max      | 58851592         | 122142375        |
| min      | 15095044         | 17856938         |
| avg.     | 40697668.4706525 | 39822508.8662606 |
| count    | 3663             | 3776             |
| stdev    | 5228744.19346652 | 8064316.19743504 |
| se       | 86393.0757542369 | 131235.567943636 |
| 95%CL    | 169327.316991945 | 257216.986660185 |
| C.L.max  | 40866995.7876444 | 40079725.8529208 |
| C.L.min  | 40528341.1536605 | 39565291.8796004 |

very small speed-up...