@MaverickTse 2016-10-31
The average is only 1ms difference while stdev is a LOT wider... so nope. 
I have also tried the Concurrency library available in VS2015 (ppl.h), also to no avail. The /Qpar is still the most stable and fastest.

If you want to continue to play with the multi-threading issue, try to parallelize the LUT lookup. So far I failed to do this and time of processing increase with number of channels being processed.