@MaverickTse 2016-11-02
Note that I'm not going to merge unless there are enough improvement in speed.