@yumetodo 2016-11-12
やっとテーブルの動的メモリー確保を削れた。
テストコードより、もとの計算結果と100%一致することを確認。

これでやっとAvUtlでの実地試験に望める。