@yumetodo 2016-11-12
Ut Video Codec SuiteなAVI素材だとなんかバグっているけど
![image](https://cloud.githubusercontent.com/assets/10869046/20237427/7e8a6c5c-a915-11e6-9048-a66db961399d.png)
とりあえず1000 micro sec.～3000 micro sec.にまで高速化できたので成果はあった。