@MaverickTse 2016-11-12
このバッグを治ったらmerge する