@yumetodo 2016-11-12
とりあえず範囲外についても元と同じ挙動にした。これよりAviUtl上でのテストに入る