@yumetodo 2016-11-12
治ったっぽい・・？
![image](https://cloud.githubusercontent.com/assets/10869046/20237670/d23349f8-a91c-11e6-932e-d276e3e6372b.png)